added **Value(Node node) Mehtods for all common xsd data types to RDFBackend; added AbstractBackend implementation to the ldpath-core that provides implementation of those based on the stringValue(Node node) Method; Addaped the GenericSesameStore to the changes in RDFBackend; added execute Method to Programm; Added Transformers for Byte, Short and Duration; Byte, Short, Integer, Long and BigInteger now support decimal values that can be converted to integers e.g. '1.0';

+++ ldpath-core/src/main/java/at/newmedialab/ldpath/model/transformers/IntTransformer.java
@@ -38,7 +38,7 @@ public class IntTransformer<Node> implements NodeTransformer<Integer,Node> {
     @Override
     public Integer transform(RDFBackend<Node> rdfBackend, Node node) throws IllegalArgumentException {
         if(rdfBackend.isLiteral(node)) {
-            return Long.valueOf(rdfBackend.longValue(node)).intValue();
+            return rdfBackend.intValue(node);
         } else {
             throw new IllegalArgumentException("cannot transform node of type "+node.getClass().getCanonicalName()+" to integer");
         }
